Remove redundant `requestLayout` when setting drawable on NavigationBar.
